add: option to use case insenstive rule
Rule is for not having to write all different variants of casing
to the exceptions list for a word.

You can just write one example for a word like h-e-l-l-o and it
will apply even if it is capitalized or all uppercased.

By default the setting is off but it can be enabled when
using createHyphenator function.